Users/qiazh/pre merge tikv bugfix (#461)
* Feature/rabitq submodule (#458)

* feat(SPANN): add official RaBitQ postings


Copilot-Session: 469bcdca-7ded-426c-baf3-ddf1b2958401

* build: track official RaBitQ as submodule


Copilot-Session: 469bcdca-7ded-426c-baf3-ddf1b2958401

* refactor(SPANN): remove legacy global RaBitQ adapter

Keep the official raw-head and posting-only RaBitQ integration as the sole supported path.


Copilot-Session: 469bcdca-7ded-426c-baf3-ddf1b2958401

* refactor(SPANN): make RaBitQ a global quantizer

Use the official scalar quantizer APIs through the standard IQuantizer workflow and remove the posting-only RaBitQ path.


Copilot-Session: 469bcdca-7ded-426c-baf3-ddf1b2958401

* clean some code

* Feature/rabitq submodule (#454) (#456)

* feat(SPANN): add official RaBitQ postings




* build: track official RaBitQ as submodule




* refactor(SPANN): remove legacy global RaBitQ adapter

Keep the official raw-head and posting-only RaBitQ integration as the sole supported path.




* refactor(SPANN): make RaBitQ a global quantizer

Use the official scalar quantizer APIs through the standard IQuantizer workflow and remove the posting-only RaBitQ path.




---------

Copilot-Session: 469bcdca-7ded-426c-baf3-ddf1b2958401

* clean some code

* remove unused code

---------

Copilot-Session: 469bcdca-7ded-426c-baf3-ddf1b2958401

* make RabitQ to be a configurable feature and fix FineGrainedLock for LargeVID

* add sequantial TPT build to save memory

---------

Copilot-Session: 469bcdca-7ded-426c-baf3-ddf1b2958401

+++ AnnService/inc/SSDServing/SSDIndex.h
@@ -198,7 +198,7 @@ namespace SPTAG {
                         return ret;
                     }
                     auto warmupQuerySet = queryReader->GetVectorSet();
-                    int warmupNumQueries = min(warmupQuerySet->Count(), p_opts.m_queryCountLimit);
+                    int warmupNumQueries = min(static_cast<int>(warmupQuerySet->Count()), p_opts.m_queryCountLimit);
 
                     std::vector<QueryResult> warmupResults(warmupNumQueries, QueryResult(NULL, max(K, internalResultNum), false));
                     std::vector<SPANN::SearchStats> warmpUpStats(warmupNumQueries);
@@ -226,7 +226,7 @@ namespace SPTAG {
                     return ret;
                 }
                 auto querySet = queryReader->GetVectorSet();
-                int numQueries = min(querySet->Count(), p_opts.m_queryCountLimit);
+                int numQueries = min(static_cast<int>(querySet->Count()), p_opts.m_queryCountLimit);
 
                 std::vector<QueryResult> results(numQueries, QueryResult(NULL, max(K, internalResultNum), false));
                 std::vector<SPANN::SearchStats> stats(numQueries);
